test(aerogpu): fix scanout handoff + VBE LFB base assertions
- Enable PCI COMMAND.BME in scanout handoff test before calling display_present

- Drop redundant u64::from casts in vbe_lfb_base tests (constants are already u64)

+++ crates/aero-machine/src/lib.rs
@@ -12547,7 +12547,7 @@ mod tests {
         .unwrap();
         assert_eq!(
             headless.vbe_lfb_base(),
-            u64::from(firmware::video::vbe::VbeDevice::LFB_BASE_DEFAULT)
+            firmware::video::vbe::VbeDevice::LFB_BASE_DEFAULT
         );
 
         // When VGA is enabled, the BIOS should report the legacy MMIO-mapped SVGA base.
